Refactor CompileTools
